refactor(dsl): move context band conflict check into its own file
